Refactor admin authentication to use React Context for state management
