+++ src/admin.tsx
@@ -1,10 +1,13 @@
 import React from 'react';
 import { createRoot } from 'react-dom/client';
 import AdminApp from './components/admin/AdminApp';
+import { AdminProvider } from './context/AdminContext';
 import './index.css';
 
 createRoot(document.getElementById('root')!).render(
   <React.StrictMode>
-    <AdminApp />
+    <AdminProvider>
+      <AdminApp />
+    </AdminProvider>
   </React.StrictMode>
 );
